requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.2.5
django-environ==0.4.5
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
psycopg2==2.8.5
pur==5.2.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.5
whitenoise==4.1.3
